requirements.txt
Flask
Flask-Cors
SQLAlchemy
alembic
PyMySQL
python-dotenv

argon2-cffi
PyJWT

cryptography